+++ .idea/encodings.xml
@@ -3,7 +3,7 @@
   <component name="Encoding" native2AsciiForPropertiesFiles="true" defaultCharsetForPropertiesFiles="GBK">
     <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/ColumnTypeEnum.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/DbException.java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/EasyDbImpl.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/EasyDbImpl1.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/Table.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/easydb/src/main/java/zyf/easydb/annotation/DbTable.java" charset="UTF-8" />
     <file url="PROJECT" charset="UTF-8" />
